--- database/factories/Accounting/EstimateFactory.php
@@ -1,23 +1,0 @@
-<?php
-
-namespace Database\Factories\Accounting;
-
-use Illuminate\Database\Eloquent\Factories\Factory;
-
-/**
- * @extends \Illuminate\Database\Eloquent\Factories\Factory<\App\Models\Accounting\Estimate>
- */
-class EstimateFactory extends Factory
-{
-    /**
-     * Define the model's default state.
-     *
-     * @return array<string, mixed>
-     */
-    public function definition(): array
-    {
-        return [
-            //
-        ];
-    }
-}
